+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Cleaned up `AbstractCommandController`
 * Improved `FileService` (breaking change)
 * Cleaned up `SlugViewHelper`